refactor(workspace): extract WorkspaceToolbar component from WorkspacePage
