+++ packages/app/src/pages/WorkspacePage.tsx
@@ -1,10 +1,7 @@
 import { useState, useEffect, useCallback, useRef, useMemo } from 'react';
-import { useNavigate } from 'react-router-dom';
 import type { ImageViewMode, DragData } from '../types/gridfinity';
 import { useWorkspace } from '../contexts/WorkspaceContext';
 import { useGridTransform } from '../hooks/useGridTransform';
-import { useUpdateLayoutMutation } from '../hooks/useLayouts';
-import { buildPayload } from '../utils/layoutHelpers';
 import { DimensionInput } from '../components/DimensionInput';
 import { GridPreview } from '../components/GridPreview';
 import { GridSummary } from '../components/GridSummary';
@@ -15,8 +12,8 @@ import { ZoomControls } from '../components/ZoomControls';
 import { ImageViewToggle } from '../components/ImageViewToggle';
 import { GridViewport } from '../components/GridViewport';
 import { SidebarPanel } from '../components/SidebarPanel';
-import { SubmissionsBadge } from '../components/admin/SubmissionsBadge';
 import { UserStlLibrarySection } from '../components/UserStlLibrarySection';
+import { WorkspaceToolbar } from '../components/WorkspaceToolbar';
 import { exportToPdf } from '../utils/exportPdf';
 
 const API_BASE_URL = (import.meta.env.VITE_API_BASE_URL as string | undefined) ?? 'http://localhost:3001/api/v1';
@@ -27,7 +24,6 @@ const LIBRARY_DEFAULT_WIDTH = 260;
 const GRIDFINITY_UNIT_MM = 42;
 
 export function WorkspacePage() {
-  const navigate = useNavigate();
   const ws = useWorkspace();
 
   const {
@@ -40,21 +36,15 @@ export function WorkspacePage() {
     copyItems, pasteItems, deleteSelected, rotateSelected, updateItemCustomization,
     bomItems,
     layoutMeta, isReadOnly,
-    handleSaveComplete,
-    drawerWidth,
-    drawerDepth,
     refImagePlacements, addRefImagePlacement, removeRefImagePlacement,
     updateRefImagePosition, updateRefImageScale, updateRefImageOpacity,
     updateRefImageRotation, toggleRefImageLock,
     referenceImagesForGrid,
     libraryItems, isLibraryLoading, isLibrariesLoading, libraryError, librariesError,
     categories, getItemById, getLibraryMeta,
-    handleSubmitClick, handleWithdrawLayout,
     handleClearAll, handleReset,
-    submitLayoutMutation, withdrawLayoutMutation,
-    submittedCountQuery,
     dialogDispatch,
-    isAuthenticated, isAdmin,
+    isAuthenticated,
     exportPdfError, setExportPdfError,
   } = ws;
 
@@ -68,44 +58,6 @@ export function WorkspacePage() {
   const [imageViewMode, setImageViewMode] = useState<ImageViewMode>(
     () => (localStorage.getItem('gridfinity-image-view-mode') as ImageViewMode) || 'ortho'
   );
-
-  const updateLayoutMutation = useUpdateLayoutMutation();
-
-  const [toast, setToast] = useState<{ visible: boolean; isError: boolean }>({
-    visible: false,
-    isError: false,
-  });
-
-  const toastTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  useEffect(() => () => {
-    if (toastTimerRef.current) clearTimeout(toastTimerRef.current);
-  }, []);
-
-  const handleDirectSave = useCallback(async () => {
-    if (!layoutMeta.id) return;
-    try {
-      const payload = buildPayload(
-        layoutMeta.name,
-        layoutMeta.description,
-        gridResult.gridX,
-        gridResult.gridY,
-        drawerWidth,
-        drawerDepth,
-        spacerConfig,
-        placedItems,
-        refImagePlacements,
-      );
-      const result = await updateLayoutMutation.mutateAsync({ id: layoutMeta.id, data: payload });
-      handleSaveComplete(result.id, result.name, result.status);
-      setToast({ visible: true, isError: false });
-      if (toastTimerRef.current) clearTimeout(toastTimerRef.current);
-      toastTimerRef.current = setTimeout(() => setToast(t => ({ ...t, visible: false })), 1500);
-    } catch {
-      setToast({ visible: true, isError: true });
-    }
-  }, [layoutMeta, gridResult, drawerWidth, drawerDepth, spacerConfig, placedItems,
-      refImagePlacements, updateLayoutMutation, handleSaveComplete]);
 
   const handleLibraryResizeStart = useCallback((e: React.MouseEvent) => {
     e.preventDefault();
@@ -353,113 +305,7 @@ export function WorkspacePage() {
           )}
         </nav>
         <div className="preview-toolbar">
-          <div className="reference-image-toolbar">
-            {isAuthenticated && (
-              <button className="layout-toolbar-btn layout-load-btn" onClick={() => navigate('/configs')} type="button">Load</button>
-            )}
-            {/* Unsaved layout */}
-            {isAuthenticated && !isReadOnly && !layoutMeta.id && (
-              <button
-                className="layout-toolbar-btn layout-save-btn"
-                onClick={() => dialogDispatch({ type: 'OPEN', dialog: 'save' })}
-                type="button"
-                disabled={placedItems.length === 0 && refImagePlacements.length === 0}
-              >
-                Save
-              </button>
-            )}
-
-            {/* Saved layout — draft or submitted */}
-            {isAuthenticated && !isReadOnly && !!layoutMeta.id && (
-              <>
-                <button
-                  className="layout-toolbar-btn"
-                  onClick={() => dialogDispatch({ type: 'OPEN', dialog: 'save' })}
-                  type="button"
-                >
-                  Save as New
-                </button>
-                <button
-                  className="layout-toolbar-btn layout-save-btn"
-                  onClick={handleDirectSave}
-                  type="button"
-                  disabled={updateLayoutMutation.isPending || (placedItems.length === 0 && refImagePlacements.length === 0)}
-                >
-                  {updateLayoutMutation.isPending ? 'Saving\u2026' : 'Save Changes'}
-                </button>
-              </>
-            )}
-
-            {/* Delivered (read-only) layout */}
-            {isAuthenticated && isReadOnly && (
-              <button
-                className="layout-toolbar-btn layout-save-btn"
-                onClick={() => dialogDispatch({ type: 'OPEN', dialog: 'save' })}
-                type="button"
-              >
-                Build from This
-              </button>
-            )}
-
-            {toast.visible && (
-              <div className={`save-toast ${toast.isError ? 'save-toast-error' : 'save-toast-success'}`}>
-                {toast.isError ? (
-                  <>
-                    <span>Save failed. Try again.</span>
-                    <button
-                      type="button"
-                      className="save-toast-dismiss"
-                      onClick={() => setToast(t => ({ ...t, visible: false }))}
-                      aria-label="Dismiss"
-                    >
-                      &times;
-                    </button>
-                  </>
-                ) : (
-                  <span>Saved!</span>
-                )}
-              </div>
-            )}
-
-            {isAuthenticated && layoutMeta.status !== 'submitted' && layoutMeta.status !== 'delivered' && (
-              <button
-                className="layout-toolbar-btn layout-submit-btn"
-                onClick={handleSubmitClick}
-                type="button"
-                disabled={submitLayoutMutation.isPending}
-              >
-                {submitLayoutMutation.isPending ? 'Submitting...' : 'Submit'}
-              </button>
-            )}
-            {isAuthenticated && layoutMeta.status === 'delivered' && (
-              <button className="layout-toolbar-btn layout-submit-btn" disabled type="button" title="This layout has been fulfilled">
-                Submit
-              </button>
-            )}
-            {isAuthenticated && layoutMeta.id && layoutMeta.status === 'submitted' && (
-              <button className="layout-toolbar-btn layout-withdraw-btn" onClick={handleWithdrawLayout} type="button" disabled={withdrawLayoutMutation.isPending}>
-                {withdrawLayoutMutation.isPending ? 'Withdrawing...' : 'Withdraw'}
-              </button>
-            )}
-            <button
-              className="layout-toolbar-btn layout-export-btn"
-              onClick={handleExportPdf}
-              type="button"
-              disabled={placedItems.length === 0}
-              title="Export layout as PDF"
-            >
-              Export PDF
-            </button>
-            {exportPdfError && (
-              <span className="export-pdf-error" role="alert">{exportPdfError}</span>
-            )}
-            {!isReadOnly && (placedItems.length > 0 || refImagePlacements.length > 0) && (
-              <button className="clear-all-button" onClick={handleClearAll}>Clear All ({placedItems.length + refImagePlacements.length})</button>
-            )}
-            {isAdmin && (
-              <SubmissionsBadge count={submittedCountQuery.data?.submitted ?? 0} onClick={() => dialogDispatch({ type: 'OPEN', dialog: 'admin' })} />
-            )}
-          </div>
+          <WorkspaceToolbar onExportPdf={handleExportPdf} exportPdfError={exportPdfError} />
           <ImageViewToggle mode={imageViewMode} onToggle={toggleImageViewMode} />
           <ZoomControls zoom={transform.zoom} onZoomIn={zoomIn} onZoomOut={zoomOut} onResetZoom={resetZoom} onFitToScreen={handleFitToScreen} />
         </div>
